requirements.txt
aiogram==3.3.0
python-dotenv==1.0.1
requests==2.31.0